+++ packages/ack-id/README.md
@@ -87,8 +87,8 @@ isControllerClaim(credential.credentialSubject)
 ### Schema Validation
 
 ```ts
-// Zod schema
-import { controllerClaimSchema } from "@agentcommercekit/ack-id/schemas/zod"
+// Zod v3 schema
+import { controllerClaimSchema } from "@agentcommercekit/ack-id/schemas/zod/v3"
 
 // Valibot schema
 import { controllerClaimSchema } from "@agentcommercekit/ack-id/schemas/valibot"
