+++ packages/ack-id/src/a2a/schemas/zod/v3.ts
