+++ packages/ack-pay/README.md
@@ -122,8 +122,8 @@ isPaymentReceiptClaim(credential.credentialSubject)
 ### Schema Validation
 
 ```ts
-// Zod schema
-import { paymentRequestSchema } from "@agentcommercekit/ack-pay/schemas/zod"
+// Zod v3 schema
+import { paymentRequestSchema } from "@agentcommercekit/ack-pay/schemas/zod/v3"
 
 // Valibot schema
 import { paymentRequestSchema } from "@agentcommercekit/ack-pay/schemas/valibot"
