--- packages/agentcommercekit/src/schemas/zod.ts
@@ -1,5 +1,0 @@
-export * from "@agentcommercekit/ack-pay/schemas/zod"
-export * from "@agentcommercekit/ack-id/schemas/zod"
-export * from "@agentcommercekit/did/schemas/zod"
-export * from "@agentcommercekit/jwt/schemas/zod"
-export * from "@agentcommercekit/vc/schemas/zod"
