+++ packages/agentcommercekit/tsdown.config.ts
@@ -4,9 +4,9 @@ export default defineConfig({
   entry: [
     "src/index.ts",
     "src/schemas/valibot.ts",
-    "src/schemas/zod.ts",
+    "src/schemas/zod/v3.ts",
     "src/a2a/index.ts",
-    "src/a2a/schemas/zod.ts",
+    "src/a2a/schemas/zod/v3.ts",
     "src/a2a/schemas/valibot.ts"
   ],
   dts: true,
