+++ packages/jwt/README.md
@@ -53,8 +53,8 @@ console.log(parsed.payload)
 The package provides schemas for validating JWT strings with Zod and Valibot:
 
 ```ts
-// Zod
-import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/zod"
+// Zod v3
+import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/zod/v3"
 
 // Valibot
 import { jwtStringSchema } from "@agentcommercekit/jwt/schemas/valibot"
