+++ packages/vc/README.md
@@ -107,8 +107,8 @@ const revoked = await isRevoked(credential)
 ### Schema Validation
 
 ```ts
-// Zod schemas
-import { credentialSchema } from "@agentcommercekit/vc/schemas/zod"
+// Zod v3 schemas
+import { credentialSchema } from "@agentcommercekit/vc/schemas/zod/v3"
 
 // Valibot schemas
 import { credentialSchema } from "@agentcommercekit/vc/schemas/valibot"
