+++ packages/vc/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsdown/config"
 
 export default defineConfig({
-  entry: ["src/index.ts", "src/schemas/zod.ts", "src/schemas/valibot.ts"],
+  entry: ["src/index.ts", "src/schemas/zod/v3.ts", "src/schemas/valibot.ts"],
   dts: true,
   silent: true
 })
